Extract runtime adapters into component crates
