+++ .gitignore
@@ -15,6 +15,8 @@ downloads/
 eggs/
 .eggs/
 lib/
+!crates/lib/
+!crates/lib/**
 lib64/
 parts/
 sdist/
