+++ Cargo.toml
@@ -1,3 +1,7 @@
+[workspace]
+members = ["crates/lib/*"]
+resolver = "2"
+
 [package]
 name = "mountaineer"
 # Bumped automatically by CI on a release
@@ -14,9 +18,9 @@ serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0"
 path-absolutize = "3.1.1"
 regex = "1.11.1"
-md5 = "0.7.0"
 notify-debouncer-full = "0.7.0"
 toml = "0.8"
+mountaineer-vite = { path = "crates/lib/vite" }
 
 log = "0.4"
 env_logger = "0.11"
@@ -34,6 +38,12 @@ rustc-hash = { version = "2.1" }
 libc = "0.2"
 winapi = { version = "0.3", features = ["processthreadsapi"] }
 tempfile = "3.16.0"
+
+[target.'cfg(unix)'.dependencies]
+mountaineer-hot-reload-fork = { path = "crates/lib/hot-reload-fork" }
+
+[target.'cfg(windows)'.dependencies]
+mountaineer-hot-reload-pool = { path = "crates/lib/hot-reload-pool" }
 
 
 [dev-dependencies]
