+++ crates/lib/hot-reload-pool/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "mountaineer-hot-reload-pool"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "2.0"
+tokio = { version = "1.43", features = ["io-util", "process"] }
+
+[dev-dependencies]
+tempfile = "3.16"
+tokio = { version = "1.43", features = ["macros", "rt"] }
