+++ crates/lib/vite/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "mountaineer-vite"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+md5 = "0.7.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tempfile = "3.16"
+thiserror = "2.0"
+tokio = { version = "1.43", features = ["net", "process", "time"] }
