+++ pyproject.toml
@@ -47,7 +47,8 @@ cache-keys = [
     { file = "mountaineer/**/*.py" },
     { file = "mountaineer/static/**/*" },
     { file = "src/**/*.rs" },
-    { file = "src/coordinator_assets/**/*" },
+    { file = "src/coordinator/**/*" },
+    { file = "crates/**/*" },
     { file = "Cargo.toml" },
     { file = "Cargo.lock" },
 ]
